Correct slab size calculation to honor fragmentation limit.
ComputeSlabSize incorrectly calculated the slab size when counting wasted bytes. Since blocks are always aligned to slab size, the slab was never grown. This change fixes the calculation by using the allocation size to determine wasted bytes and grows the slab in multiples of the block size if the slab size wasn't specified.

Also,
* Move SlabOnBuddy tests to SlabMemoryAllocatorTests.
* Add ResourceFragmentationLimit option to ALLOCATOR_DESC.