Remove Usage of std::set in ResidencySet
Removes usage of std::set in ResidencySet and replaces it with a vector.
This does allow duplicat heaps to be inserted into the ResidencySet,
however insertion will be O(1) instead of O(log(n)) and instead we can
easily identify duplicate heaps by checking their last used fence value
when when they are iterated.